Remove `BuiltinLintDiag::UnusedImports` variant
